fix: emulatorId 타입 수정에 따른 코드 수정

+++ back/emulator-server/src/main/java/com/example/emulatorserver/device/domain/emulator/EmulatorReader.java
@@ -6,5 +6,5 @@ import java.util.Optional;
 public interface EmulatorReader {
     Optional<EmulatorEntity> findById(int id);
     List<EmulatorEntity> findAll();
-    EmulatorEntity getById(Long id);
+    EmulatorEntity getById(int id);
 }
